Fix method sinature of FuseAppLoop JNI code
The file size passed to replyLookup and replyGetAttr needed to be jlong,
but they were jint before the fix.

Bug: 63543158
Test: Transfer 4GB+ files through ProxyFileDescriptor

(cherry picked from commit d2229a65f53e7b9e9fc861ac300607a329153c4c)

+++ core/jni/com_android_internal_os_FuseAppLoop.cpp
@@ -142,14 +142,14 @@ void com_android_internal_os_FuseAppLoop_replyOpen(
 }
 
 void com_android_internal_os_FuseAppLoop_replyLookup(
-        JNIEnv* env, jobject self, jlong ptr, jlong unique, jlong inode, jint size) {
+        JNIEnv* env, jobject self, jlong ptr, jlong unique, jlong inode, jlong size) {
     if (!reinterpret_cast<fuse::FuseAppLoop*>(ptr)->ReplyLookup(unique, inode, size)) {
         reinterpret_cast<fuse::FuseAppLoop*>(ptr)->Break();
     }
 }
 
 void com_android_internal_os_FuseAppLoop_replyGetAttr(
-        JNIEnv* env, jobject self, jlong ptr, jlong unique, jlong inode, jint size) {
+        JNIEnv* env, jobject self, jlong ptr, jlong unique, jlong inode, jlong size) {
     if (!reinterpret_cast<fuse::FuseAppLoop*>(ptr)->ReplyGetAttr(
             unique, inode, size, S_IFREG | 0777)) {
         reinterpret_cast<fuse::FuseAppLoop*>(ptr)->Break();
